requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.3
langchain-openai==0.2.0 
langgraph==0.2.20
langchain-core==0.3.0
langsmith==0.1.117
langgraph-checkpoint==1.0.14
uuid==1.30
requests==2.31.0